requirements.txt
altair==6.0.0
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
eventlet==0.40.4
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.16.0
netifaces==0.10.6
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.5
pyarrow==23.0.0
pydeck==0.9.1
PyMySQL==1.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==3.3
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
wsproto==1.3.2
